Refactor authentication imports and consolidate auth module structure
