+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,2 @@
-import { handlers } from '@/app/auth/auth'; // Referring to the auth.ts we just created
+import { handlers } from '@/lib/auth'; // Referring to the auth.ts we just created
 export const { GET, POST } = handlers;
