--- src/app/auth/index.ts
@@ -1,1 +1,0 @@
-export { handlers, signIn, signOut, auth, getUserSession } from './auth';
