+++ src/lib/auth/index.ts
@@ -1,0 +1,8 @@
+export {
+  handlers,
+  signIn,
+  signOut,
+  auth,
+  getUserSession,
+  providerMap,
+} from './auth';
